+++ src/App.js
@@ -1,19 +1,24 @@
+import { useState } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Main from './pages/Main';
 import LogIn from './pages/LogIn';
+import Application from './pages/Application';
 import Introduction from './pages/Introduction';
-import { useState } from 'react';
 
 function App() {
   const [isLoggedin, setIsLoggedin] = useState(false);
   return (
     <BrowserRouter>
       <Routes>
+        <Route index path="/" element={<Main />} />
         <Route path="/introduces" element={<Introduction />} />
         <Route
-          path="/"
+          path="/applications"
           element={
-            <Main isLoggedin={isLoggedin} setIsLoggedin={setIsLoggedin} />
+            <Application
+              isLoggedin={isLoggedin}
+              setIsLoggedin={setIsLoggedin}
+            />
           }
         />
         <Route
